% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bibliography}
\alias{bibliography}
\title{Bibliography}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 131 rows and 2 columns.
}
\usage{
bibliography
}
\description{
Reference list for marine mammal life history data
}
\details{
\describe{
\item{reference} {reference id}
\item{reference.text} {reference text}
}
}
\keyword{datasets}
