% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create marinesurv obj.R
\name{create_marinesurvival_modinput}
\alias{create_marinesurvival_modinput}
\title{Function converting  marine.lifehist.data to input for a marine survival model}
\usage{
create_marinesurvival_modinput(datasets.list)
}
\description{
Function converting  marine.lifehist.data to input for a marine survival model
}
