% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create marinesurv obj _helpers.R
\name{get_bias.direction.vector}
\alias{get_bias.direction.vector}
\title{Helper Function to sort sampling bias data}
\usage{
get_bias.direction.vector(datasetskey.df, extras.df)
}
\description{
Helper Function to sort sampling bias data
}
