% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create marinesurv obj _helpers.R
\name{get_bias.inclusion.vector}
\alias{get_bias.inclusion.vector}
\title{Helper Function to sort sampling bias data}
\usage{
get_bias.inclusion.vector(biasmat)
}
\description{
Helper Function to sort sampling bias data
}
