% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create marinesurv obj _helpers.R
\name{get_biasmat}
\alias{get_biasmat}
\title{Helper Function to sort sampling bias data}
\usage{
get_biasmat(datasetskey.df, extras.df, nages = 100)
}
\description{
Helper Function to sort sampling bias data
}
