% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fn.R
\name{get_lifehist_data}
\alias{get_lifehist_data}
\title{Filter subsets of marine.lifehist.data}
\usage{
get_lifehist_data(
  data.type = "age-structure",
  species = NULL,
  sex = NULL,
  return.key = FALSE
)
}
\arguments{
\item{data.type}{"age-structure" (default), "corpora" or "pregnancy"}

\item{species}{vector of species names, as common name in camel-case with no spacies.
If \code{NULL} (default) then all species are returned}

\item{sex}{"F" or "M", character, indiciating sexes of datasets return if
\code{NULL} (default) both sexes are returned.}

\item{return.key}{if \code{FALSE} (default) then only datasets are returned. Else
if \code{TRUE} then a filtereded data key is also returned.}
}
\value{
if \code{return.key = FALSE} then a list of datasets. A subset of \code{\link{marine.lifehist.data}}
matching the input conditions. List element names reflect dataset number

if \code{return.key = TRUE} then a two element list. Element 1 \code{$key} is a filtered
version of  code{\link{marine.lifehist.datakey}} describing the datasets matrching the input conditions.
Element 2 \code{data} is a list of datasets.  A subset of \code{\link{marine.lifehist.data}}
matching the input conditions. List element names reflect dataset number
#'
}
\description{
Function to extract subsets of \code{\link{marine.lifehist.data}} easily.
}
\details{
All data are present in \code{\link{marine.lifehist.data}}and \code{\link{marine.lifehist.datakey}}
this function simply provides a simple way to access subsets of the data. See
\code{?marine.lifehist.data} & \code{?marine.lifehist.datakey} for full description
of column headings and contents.
}
\examples{
get_lifehist_data()
get_lifehist_data(species = "KillerWhale", sex = "F")
get_lifehist_data(species = c("KillerWhale", "ShortFinnedPilotWhale"))

}
