% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine.lifehist.data}
\alias{marine.lifehist.data}
\title{Life history data from marine mammals}
\format{
`marine.lifehist.data`
A list of length 312. Where each list item is a marine life history dataset. Each
list item is a data frame with size and format depending on data type.

For \strong{age-structure data}. A data frame with two columns, where each row represents a single animal:
\describe{
\item{age}{age of animal in years}
\item{dataset}{dataset id, see \code{\link{marine.lifehist.datakey}} }
}

For \strong{corpora data}. A data frame with three columns, where each row represents a single animal:
\describe{
\item{age}{age of animal in years}
\item{corpora}{number of corpora reported from the ovary. this includes all types of corpora reported}
\item{dataset}{dataset id, \code{\link{marine.lifehist.datakey}} }
}

For \strong{pregnancy data}. A data frame with three columns, where age row represents an age in the sample:
\describe{
\item{age}{age in years}
\item{pregnancy.rate}{proportion of females of a given age pregnant in sample}
\item{dataset}{dataset id, \code{\link{marine.lifehist.datakey}} }
}
}
\usage{
marine.lifehist.data
}
\description{
List of life history data extracted from the literature for marine mammals.
Details of data types, species, and references are in \code{\link{marine.lifehist.datakey}}
}
\keyword{datasets}
