% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine.lifehist.datakey}
\alias{marine.lifehist.datakey}
\title{Key to data listed in marine.lifehist.data}
\format{
`marine.lifehist.datakey`
A data frame with 312 rows and 9 columns. Each row corresponds to an item in
\code{\link{marine.lifehist.data}}
\describe{
  \item{dataset}{numbered dataset id}
  \item{species}{species name: camel-case common name with no spaces. For other names see \code{\link{marine.lifehist.speciesdata}}}
  \item{sex}{sex of whales in dataset}
  \item{reference}{reference id for data origin. see \code{\link{bibliography}} for full reference list}
  \item{data source}{whaling, mass-stranding, bycatch etc.}
  \item{from} {table, barplot, scatterplot etc.}
  \item{data} {Data type, currently: age-structure, corpora or pregnancy}
}
}
\usage{
marine.lifehist.datakey
}
\description{
Data key to marine.lifehist.data
}
\keyword{datasets}
