% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine.lifehist.datasetextras}
\alias{marine.lifehist.datasetextras}
\title{Extra information on datasets included in marine.lifehist.data}
\format{
`marine.lifehist.datasetextras`
A list of length 2. Each item is a data frame describing different species level data.
List items are ordered and named:

\enumerate{
\item{\emph{datasetsextras_bias.details}} data frame indicating known sources of bias (metric) and their (value) needed for survival calculations
\item{\emph{datasetsextras_populations.key}} data frame showing which datasets are taken from studies of the same population (pop.identifier)
}
}
\usage{
marine.lifehist.datasetextras
}
\description{
Some further information on datasets in \code{\link{marine.lifehist.data}} necessarry
for some types of analysis
}
\keyword{datasets}
