% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine.lifehist.speciesdata}
\alias{marine.lifehist.speciesdata}
\title{Species level data for Marine Mammals}
\format{
`marine.lifehist.speciesdata`
A list of length 2. Each item is a data frame describing different species level data.
List items are ordered and named:

\enumerate{
\item{\emph{species_names.key}} data frame with species names and alternatives, as per International Whaling Commission Sept 2021
\item{\emph{species_age.maturity}} data frame with species ages at maturity for each sex of each species. NAs indicate no data found
}
}
\usage{
marine.lifehist.speciesdata
}
\description{
List of different species level data for marine mammals.
}
\keyword{datasets}
